rollback unsafe optimization and fix some
There are many places in the code where someone complicated
object used only by a single function is made 'static const'
within the function body.  The intent is to avoid the overhead
of initializing the object many times.  This sounds good, but
is completely not thread-safe in many cases.

Since it's so easy to realize the same savings in an actually
thread-safe manner, we should just do that.  There are lots of
ways to do this, including moving your 'static const' variable
to the global scope or to an instance variable that gets constructed
at object creation time.  I've fixed many of them this way.

Still, there are more that are not-so-trivial to fix or "not
worth my time right now", so I've filed SH-8 to track them and
marked them in the code.  If anyone what's to knock off SH-8
while they're waiting for a pizza or something, please by my
guest.

There was one instance of this trick which *was* causing the
game to blow up because the code *is* accessed from multiple
threads simultanously.  See the change in csg::region.cpp for
that one.