Final fix to the camera dragging; turns out it was a precision issue.
Originally, while dragging, every mouse-move would trigger a new translation of the camera.  This translation was a delta between the newly calculated position, and the last-calculated position.  There's a fair bit of floating point math in there, too.  Dragging your mouse across the screen causes dozens (maybe low hundred) of these events to fire, so hundreds of little FP operations would occur.  And so, lots of precision issues would result, and in the end, your cursor would drift from your selection.

Now, compute the delta between the start and the current pixel, and change the camera integration scheme a bit to handle it.