clean up the data directory + cursor handling
Remove a bunch of dead code from the data directory.

Also begin moving the cursor handling code to lua.  This adds
the client call _radiant.client.set_cursor to create a cursor
and add it to the top of the stack.  Use :destroy() on
the returned promise to remove a cursor from the stack.  Create
stockpile shows how this is done.  Eventually, all cursor handling
will be done in lua, so go ahead and move all the cursors in there
now.

The only directory left hanging out in stonehearth_data now is
horde.  Yay!

I also fixed the implementation of AddTraceRenderFrameHandler.
The current implementation would fail to call subsequent handlers
if any previous handler returned false.  The new one calls all
handlers unconditionally.