Use the random number geneartor library almost everywhere
We had a bug in one site that was using naked rand() where if the number chosen was exactly RAND_MAX we would overflow an array (technically, the ASSERT protecting us from overflowing the array would fire, but the effect was the same: bad things man, bad things)..
Fix that, and more generally use the random number generator library (which gets edge cases like this correct in a much more friendly interface) in places which make sense.  There are still some embedded deep in horde that i'm leaving alone for now.