Break the NavGrid tile up into two classes to save memory.
NavGridTile stores the list of entities colliding with the tile.  A linked NavGridTileData class stores the actual bitvectors.
Finally, the NavGrid manages an LRU list of NavGridTiles and throws out their data when the working set gets to big.
The current default is 128 tiles.  While profiling a release build, I discovered we were allocating over 100 MB in NavGridTiles,
99% of which were never referenced (e.g. trees off in far corners of the world).  This should definitely help.