fix "pick a random personality"
The current code had an edge condition where it would enter an infinite loop.  Changed the implementation to be more efficient (and avoid the infinite spin).
Now the running time is O(1) instead of O(time it takes a random number to not hit a false boolean) (which I think is O(infinite), technically =)