grealy improve pathfinder performance.
Two big changes here.  The first is to eliminate some of the std::unordered_map's used for storing
f, g, and h for A*.  Instead, just store the variables in the nodes contained in the open list.  This greatly
reduces the constant factor on lookup of the nodes.  I have intentionally stored value types rather than
pointer types in the open list to improve cache coherency.

The second is to group all jobs for an entity into a master EntityJobScheduler class which divides the
job time slice bewteen the jobs most likely to succeed.  This grealy reduces the "140 pathfinders all
running in parallel" problem.  Now the number of pathfinders running in parallel is more in proportion
to the number of entities in the world which are trying to get somewhere.