Stop using faction for everything.
"faction" is used to determine alliances.  It's weird that we use it for
so much other stuff (eg. identifying all the units belonging to a
particular player).  This change adds the "player_id" concept to tag
things belonging to a particular player and puts both that and the
"kingdom" for that player in the session object so call_handlers can
use it to get to the right data.

For more info, see: http://wiki.rad-ent.com/doku.php?id=player_model