take extra care when restocking to not put items on top of each other.
There was a very very narrow race between a terrain trace firing to let the stockpile
know an item had been added and all the other restock tasks in the system all
trying to find a path to the stockpile.  If the wrong person "lost", the reservation on
the stockpile destination of the first drop would be released, a pathfinder would find
that exact location for another drop, and *then* the terrain trace callback would fire,
adding that destination to the region.  Close that window by adding the drop point
to the region immediately after the drop in  RestockStockpile.