Remove the one and ten minute poll.
The poll events in events and calendar are undesirable, as they stack a TON
of work at a single game instant.  This *could* result in big freezes very now
and then, so their use should be discouraged.

Replace use of the one and ten minute polls in the analytics service with
a real time timer, then remove the one and ten minute polls themselves (so no one
adds another dependency)