add a cost heuristic to the ai system
actions can now call ai:set_cost(n) to indicate how much an action
costs.  the units of "cost" are the time it takes an unladen person
to walk 1 tile, which is nice as people who need to walk places can
just put in the distance to that.  if 2 action with equal priority
simultaneously become ready (which happens ALL THE TIME now that
we have a direct path finder), the one that's cheapest wins.

Add ai:set_cost() calls to the follow path action.  Yay!  No farmers
aren't so crazy!