grand unification of models and such.
Non-animated models are now loaded and meshed in the coordinate system of the file.  This means vertex coordinates are all integers and the collision shapes derived from voxel meshes will now be nearly identical.  No more magic +0.5!
The mob component now has "local_origin" and "render_offset" parameters to identify the position in the mesh to rotate around and how far to tweak the rotated mesh when rendering, usually to get it to align to a grid.