allow an action to serve as a filter to prevent the action from being running.
steps:
1) put YouAction.type = 'filter' in your header.
2) Implement :should_start_thinking() which returns true if you think the action should run and false otherwise.  May not be asynchronous.

The AI framework will look at all the actions for the activity.  Before doing anything, it will take the ones marked as filter and call 'should_start_thinking' on them.  If any of them return false, the frame stops.
Everything that's left (i.e NOT the filters) participates in the election as before.