stop record change traces from spamming.
if a sync and async trace were placed on the same record, the sync
trace would fire multiple times: once when a record field changes and
again when the tracer for the async trace flushed.  this happened because
the implementation of "notice that one of my fields has changed" both
pushed the change bit into the tracer *and* notified the trace.  notifying
the trace for the sync record traces would call the callback!

so now, just notify the trace immediately when the field changes and
*only* push the change bits into the tracer.