store, meet streamer.  streamer, meet store!
Introduce the store and the streamer to each other, and make them
cooperate during encoding.  Before, they would only communicate  through
the StoreTrace interface, which meant the Streamer had to put a trace on
every object in the world (including all the record fields).  This led to
great sadness, as traces are expensive.

A trace's value (for the streamer) is primarily in its ability to buffer
delta state for containers before pushing it over the wire.  So let's keep
those around.  For all other objects, though (Boxes and Records,
basically) have the store notify the streamer DIRECTLY when an object
changes.

For a Quick-Start New Game, this drops the number of traces created from
around 830,000 to around 180,000.  Not bad!