f*#$&(*ing floats, man.
The maximum value of a float such that (int)(((float)x) + 1) == x + 1
for integer x is around a million.

Yeah.

This is causing lots of off-by-one errors in the lua side (remember, all
numbers are floats now in lua), especially in constructing walls, etc.

So standardize on doubles.  It's this or the fixed-point insanity (LOL!)