better iterators for regions (see below).
So, code which used to do this:

for (auto cube : region)
  for (auto pt : cube)
     // Do something assuming 'pt' is visisted exactly once

Was wrong for all floating point regions (to see why, consider what happens when 2 cubes in a region share a non-integer aligned edge).

To correct that, all the iterators for cubes and regions have change!

1) You must explicitly call EachCube() or EachPoint() to iterate over all cube or points in a region (respectively).  Cube and Points returned by these iterators are guaranteed to be unique
2) You must explicitly call EachPoint() to iterate over all points in a cube (since symmetry is nice).
3) "for (auto cube : EachCube(r)) for (auto pt : EachPoint(cube))" is still valid, but know you might get non-unique points!

Lua gets :each_cube() and :each_point() and must obey the same rules.

This fixes many crashing or infinite loop bugs (e.g. try placing a tombstone a few times.  BOOM).