better memory stat logging.
log the address space info as well as the amount of memory used.
the actual upper-bound of the amount of memory we can allocate is
determined by the virtual address space, which is machine dependant.
On mine it happens to be 1.3 GB, but on another machine with a different
PCI layout or different drivers it might be much bigger or much smaller.

Here's a sample of the new log after a crash where I allocated 1
MB at a time in a tight loop:

2014-Oct-21 16:05:34.122833 Total System Memory:     31.875 GB (34225917952 bytes)
2014-Oct-21 16:05:34.124834 Current Memory Usage:    1.374 GB (1475149824 bytes)
2014-Oct-21 16:05:34.126836 Total Address Space:     2.000 GB (2147352576 bytes)
2014-Oct-21 16:05:34.128838 Available Address Space: 5.234 MB (5488640 bytes)
2014-Oct-21 16:05:34.130840 Used Address Space:      1.995 GB (2141863936 bytes)

"Available Address Space" is the interesting bit.  If that's really small,
we've run out of memory.  Fragmentation may still be a problem, but this is
strictly better, so let's go with it.