don't run too far head of the client
Send an ack from the client to the server everytime it finishes processing an update.
If the client gets 10 updates ahead of the server, stop encoding game state and stuff.  The server can still respond to requests (e.g. pushing trace data and remote calls), but it will stop encoding the state of all the data model objects from the streamer.  The changes get buffered up in traces and will be flushed as soon as the client eventually "acks".