demand_tribute partial implementation and bulletin enhancements
added the basic state machine for the intro and shakedown portions
of the derive_tribute encounter.

added a keep_open flag to keep a dialog box open even ever the user
clicks a button.

made sure bulletins rendered and behaved properly when their user data
changes.  with this and the above change, you can now use a bulletin to
implement a paged dialog, a simple conversation tree, etc.

make sure the bulletin view and dialog are destroyed if they are removed
from the list by the server.