use the interval passed by the server for client interpolation
and fix the interval passed by the server, which is why the
client thought it had to figure this out to begin with!

also bonus logging which helped me find this bug.

this fixes the "feet skating" problem after load and other timing
related bugs where the server timestamp changed wildly (e.g after
load).