fix mouse offset problems in the browser.
we weren't updating the browser offset when the window
changed, but the size of the screen didn't (e.g. take
a window with black borders and resize so the borders
get a little wider)

also change the Rect2 passed in OnScreenResize to an
actual rect instead of an offset/size tuple.