opt-into slow start based on the activity instead of opt-out.
slow-start is the mechanism we use to avoid calling start-thinking on
a very large number of execution frames, only to have 1 of them complete
very quickly (wasting the time we spent thinking all the others!).

this is a huge performance win when there are a large number of units
in a given frame.  it's an even bigger win when the act of calling start_thinking
on those units is very expensive (e.g. starting pathfinders).  this happens a
lot during mining and building.

where it's not a win, we do not want to slow-start, as it has some strange
side-effects.  if there are many units which complete immediately, they
will be "starved out" by the unit which last run (since all the others are
slow-started).  if they could have completed immediate and have a higher priority
than the rerun unit, you get this weird behavior where the rerun unit gets to
go for 250 ms, then gets immediately pre-empted by the higher priority one.
this happens a lot during idle (starvation) and combat (pre-emption).

also, generally, everyone in the execution frame needs be playing by the
same rules.  we don't want to get into a situation where 1/2 the units in a
frame slow start and the other 1/2 fast start.  that's just weird.

this commit changes the mechanism from opt-out to opt-into slow start, and
puts the config option on the activity rather than the action.  this ensures
that all EU's in and EF are fast or slow started and defaults to non-spooky
behavior over performant-behavior.  it's fairly trivial to find the things
which need to slow-start: it's the code which creates tons of tasks on an
activity name.

Actions can still explicitly opt-out of slow-start using .realtime (since
that code path already worked), but it's really really really not recommended.