fix lots of boundary conditions around grow roof
- connect roofs directly to walls and columns rather than looking
upward to try to find the roof.  this fixes weird edge cases where
a wall under the overhang of the roof would grow just it's leftmost
edge all the way up.

- only allow at most 1 roof for a wall or column rather than 1 roof
per building. this is useful when making U shaped buildings with a
roof at the top of each spire.

- better preview in the grow roof tool when you can't grow a roof above
the hovered wall.