remove bogus _starting check.
It's not clear what this was trying to do when it was written
way back then, but this is clearly bogus.  If we ever abort
in :start(), there's no way to start thinking!  I'm guessing
this code was written at a time when the EU and EF would die
on abort.