better searching code.
starting from a point on the convex hull, try to reach
a point `d` units away that is both theoretically reachable (*)
and can contain the client specified reserved region.

(*) one shortcut taken here is to only testing to see if
a single point is standable rather than an entity sized object.
this speeds up the search, but can lead to inaccuracies.