simply the pathfinding tasks for fabrication.
before this change, we would find a path to a fabricator
to figure out what nearby reachable blocks are made of, then
bfs to pick up that material, then find a path BACK to the block
we found original.  that's a lot of pathfinding to come up with
non-optimial solutions and genearlly was just... bad.

now we break the fabricator region up into several proxy entities
each of which represents a single material.  now we can path find
directly to a material followed by a path find directly to the
proxy.  less work and more optimal solutions!