Updating hud nodes was hilariously expensive (constantly dirtying all ancestors, constantly updating buffers), so stop doing both of those things.
We only needed to update buffers because we weren't using shaders to generate vertex positions, and we only needed the dirtying because we needed to update buffers.  We're not updating the hud element bounds quite correctly anymore, but I haven't seen any artifacts yet.  If you see disappearing overlays, let me know!