Various deferred renderer changes greatly improving performance w/ lots of lights.
Go back to rendering deferred lights with spheres instead of the projected screen-space bounding box of the world space bounding box of the light's sphere.  Hurr.
Revert to a 32bpp framebuffer for light accumulation.  We still saturate in the forward renderer (I'm not anxious to make another pass).
Include a damn depthbuffer when drawing lights so we can avoid overdraw!