Completely re-jigger selection.
Selection no longer sets a bit on the model; instead, the selected nodes (the logical root of the whatever model is selected) is stored, and the DrawSelected pipeline command now explicitly draws those models.
Colors can now be arbitrarily set for a selected material, via lua as well.
Selection drawing no longer queries the entire spatial graph; the selection shader discards via alpha test, so fewer texture fetches are done.
Scene now exposes a way to specifically enqueue a single node (and all its children) for rendering.
The 'selection bit stuck when object is dropped' is, as a consequence, fixed.