save bits directly into the saveid instead of tmpdir
steam users are reporting weird issues where we do not have
permission to move the tmpdir after dumping all the files into
it.  maybe the steam client is doing something in there?  anyway,
it's fairly trivial to just drop the files in the savedir, so
go back to doing that.

this also re-implements override as "save and delete the old thing
if that save was successful" so we don't lose saves if the overwrite
fails.