Completely re-jigger the way nodes are updated, including a lot more correctness and asserts, but also a bit more complexity :-(.
The form is right for performance (multiple queues), but it's this mismash of grid hashing and flags (NoCull, NoDraw), as well as possible simultaneous updates to flags and position, that make things touchy and confusing.

A better solution might be to map the hash space into somewhere smaller than 32 bits, and use the remaining flags for the NoCull/NoDraw queues.  I'll get to that soon-ish, I hope.