add the carry monitor to euv2 and use it.
when the carrying changes deep in an inner node of the ai
tree, we need to stop thinking immediately.  relying on the
restart_thinking from top won't save us, as that restart will
at best restart all the siblings of the topmost running node.
in fact, EVERY node in the tree needs to stop if they care,
especially the 3rd and 4th nephews of some currently active
top node (this happens a LOT with pickup)