turn off the sampling profiler by default.
So, it turns out luaV_execute is very reentrant.  This makes
it REALLY HARD to discount time spent not running the lua
interpreter (e.g. pathfinding).

We could zero out the delta time at the top of luaV_execute
(which is what we do now), but that loses tons of time when
the recursive step is taken.

We could try finding every recursive call to luaV_execute and
accmulating the delta before the call (while leaving the zeroing
in), but that's JUST STUPIDITY.

We could try counting the number of times we call luaV_execute,
but I'm pretty sure lua will long jump way down the stack on
error, so our count will always be wrong.

So, I can't figure out a good way. Until this is fixed, nearly all
the time spent in lua gets tossed.  So switch the default back to the
line counting version.