Single-step debug individual pathfinders!
Typical usage: hit F3 (to disable all pathfinders).  Do whatever it is that causes a pathfinder to be created (e.g. request an object be moved).  Turn on AI inspector.  Select that entity.  Find the pathfinder that you wish to step; its search progress text should be green (if not, wait a little).  Hit pause (otherwise thinking will restart over and over).  The 'step pathfinder' link at the top should now be enabled and with the job id of the pathfinder you want to step.  Click it, to step that pathfinder.