Update cubemitter 'angle' parameter to function more intuitively.
The parameter now defines the _set_ of angles from which to emit new cubes.  A 'CONSTANT' type means the cubes are emitted at exactly one angle around emitter (so a value of 0 would be all cubes in a straight line; a value of 45 would from a cone; a value of 90 would form a plane).  A 'RANDOM_BETWEEN' chooses any angle between the supplied values: [0, 0] is again a straight line, [0, 45] defines a filled cone, [0, 180] will emit cubes in every direction.